Extensions

Originally, GeMM was designed to model island plant communities, but it has since been adapted to various other study questions. These files include functions that are not needed by the "core" model, but were added for these other studies.

habitatchange.jl

Investigate the effect of temporal habitat change on plant communities.

GeMM.changehabitat! — Method

changehabitat!(world, settings)

Carry out 'global change' on all patches.

GeMM.changeprec! — Method

changeprec!(world, sdprec)

Change the precipitation of all patches according to a normal distribution.

GeMM.changetemp! — Method

changetemp!(world, sdtemp)

Change the temperature of all patches according to a normal distribution.

GeMM.markthem! — Method

markthem!(community)

Set each individual in the community (= array of individuals) as "marked".

GeMM.markthem! — Method

markthem!(world)

Set every individual in the world as "marked".

GeMM.markthem! — Method

markthem!(habitat)

Set each individual in the given patch as "marked".

invasion.jl

Carry out species invasions experiments in an island setting.

<!–FIXME Documenter.jl doesn't recognise function definitions that are not on the top level - like the invade!() functions. –>

GeMM.createind — Function

createind(settings, marked=false)

Create an individual organism of a new species with a random genome.

GeMM.disturb! — Method

disturb!(world, settings)

Disturb all patches in the world.

GeMM.disturb! — Method

disturb!(patch, intensity)

Species-independent mortality due to disturbance on the given patch. The intensity gives the mortality percentage.

zosterops.jl

Adapt GeMM to investigate eco-evolutionary population dynamics of East-African Zosterops populations in the Taita Hills, Kenya.

GeMM.zdisperse! — Function

zdisperse!(world, settings)

Dispersal of bird individuals within the world. Males disperse first, looking for suitable habitats within their dispersal range to establish territories. Females disperse second, looking for available mates. (Cf. Aben et al. 2016)

GeMM.zdisperse! — Method

zdisperse!(bird, world, location, cellsize)

Dispersal of a single bird.

GeMM.ziscompatible — Method

iscompatible(female, male, tolerance)

Check to see whether two birds are reproductively compatible.

GeMM.zreproduce! — Method

zreproduce!(patch, settings)

Reproduction of Zosterops breeding pairs in a patch.