Model Framework

These are the core functions of the model, implementing the high-level structure of the simulation and the object classes that are involved.

entities.jl

GeMM.AbstractGene — Type

An abstraction layer to hide the implementation details of genes.

GeMM.BigGene — Type

A variation of the standard Gene struct that allows for longer base sequence lengths through its use of BigInt.

GeMM.Chromosome — Type

A struct that saves a collection of genes and a tag to denote whether they are maternal or not.

GeMM.Gene — Type

The standard gene struct; including the base sequence (represented in binary and stored and as an integer), and an array of traits the gene codes for.

GeMM.Individual — Type

One of the core structs of the model, representing an individual organism.

GeMM.Patch — Type

One of the core structs of the model, representing a one-hectare patch of ground.

GeMM.Trait — Type

A simple struct to store organism trait names and values.

initialisation.jl

GeMM.createpop — Method

createpop(settings)

Create a new, random individual and replicates it a certain number of times (depending on metabolic variables) to create a new population of organisms. Returns an array of individuals.

GeMM.createworld — Method

createworld(maptable, settings)

Use a parsed map file (as returned by readmapfile) to create the world. Initialises each patch with its parameters and a new community, then returns an array of patches.

GeMM.genesis — Method

genesis(settings)

Create a new community, composed of random new species populations, for a patch. Returns an array of individuals.

GeMM.updateworld! — Method

updateworld!(world, maptable, cellsize)

Reinitialise the world from another parsed map file. Works analogously to createworld. Intended for use in scenarios where the model world changes during a run (e.g. through global warming or island ontogeny).

GeMM.varyalleles! — Method

varyalleles!(genes, settings, locivar)

Mutate gene traits in the passed array of genes.

GeMM.varyalleles! — Method

varyalleles!(chromosomes, settings, locivar)

Mutate gene traits in the passed array of chromosomes.

scheduling.jl

GeMM.defaultexperiment — Method

defaultexperiment(world, settings)

The standard annual update procedure, designed primarily for plant communities.

GeMM.invasionexperiment — Method

invasionexperiment(world, settings)

The annual update procedure for the invasion experiments.

GeMM.simulate! — Function

simulate!(world, settings, timesteps)

This is the central function of the model with the main event loop. It defines the scheduling for all submodels and output functions.

GeMM.zosteropsexperiment — Method

zosteropsexperiment(world, settings)

The annual update procedure for the Zosterops experiments, this time for bird populations.