Input, Output, and Settings

These functions are responsible for reading in all model configurations (passed by config file or commandline), administrating them during a run, and printing any output.

constants.jl

This file defines constants that are needed for calculations involving the metabolic theory of ecology.

Currently includes boltz (the Boltzmann constant) and act (the activation energy).

defaults.jl

GeMM.defaultSettings — Method

defaultSettings()

Defines the list of configuration variables and returns their default values in a Dict.

input.jl

GeMM.basicparser — Method

basicparser(filename)

Do elementary parsing on a config or map file.

Reads in the file, strips whole-line and inline comments and separates lines by whitespace. Returns a 2d array representing the tokens in each line.

GeMM.getsettings — Function

getsettings()

Combines all configuration options to produce a single settings dict. Order of precedence: commandline parameters - config file - default values

GeMM.parsecommandline — Method

parsecommandline()

Certain parameters can be set via the commandline.

GeMM.parseconfig — Method

parseconfig(filename)

Parse a configuration file and return a settings dict.

The config syntax is very simple: each line consists of a parameter name and a value (unquoted), e.g. nniches 2. # is the comment character.

GeMM.readmapfile — Method

readmapfile(mapfilename, settings)

Parse a map file and return the number of timesteps this map is to be used for (first line of the file) and the patch definitions. The latter is used by createworld and updateworld!.

output.jl

GeMM.dumpinds — Function

dumpinds(world, io, sep)

Output all data of individuals in world as table to io. Columns are separated by sep. WARNING: this produces very large files!

GeMM.makefasta — Function

makefasta(world, settings, io, onlyisland, sep)

Record the genome of every individual currently alive to the given IO stream. WARNING: this produces very large files!

GeMM.printheader — Function

printheader(settings, io, sep)

Print a list of property names to the given IO stream. This is a helper function for dumpinds.

GeMM.printpopheader — Method

printpopheader(io)

Print a list of property names to the given IO stream. This is a helper function for printpopstats.

GeMM.printpopstats — Method

printpopstats(io, world, settings, timestep)

Record statistical information (maximum, minimum, median, standard deviation) for a range of individual properties, as seen over the whole world population.

GeMM.recordlineages — Method

recordlineages(w)

Save the abundance of each lineage per patch.

GeMM.recordstatistics — Method

recordstatistics(w)

Write out world properties to the log file for later analysis.

GeMM.setupdatadir — Method

setupdatadir(dir)

Creates the output directory and copies relevant files into it.

GeMM.simlog — Function

simlog(msg, settings, category, logfile, onlylog)

Write a log message to STDOUT/STDERR and the specified logfile (if logging is turned on in the settings).

Categories: d (debug), i (information, default), w (warn), e (error)

If logfile is the empty string (default: "simulation.log"), the message will only be printed to the screen. If onlylog is true (default: false), the message is not printed to screen but only to the log.

GeMM.writedata — Method

writedata(world, settings, timestep)

Writes simulation output from world to separate table and fasta files. (Which data is recorded depends on the settings.) timestep and setting information is used for file name creation.

GeMM.writesettings — Method

writesettings(settings)

Record the settings actually used for a simulation run (cf. getsettings). Creates a config file that can be used for future replicate runs. Also records a time stamp and the current git commit.