Main Model

Main Model

These are the core functions of the model, implementing the high-level processes of the simulation and the object classes that are involved.

entities.jl

GeMM.Patch — Type.

One of the core structs of the model, representing a one-hectare patch of ground.

source

GeMM.AbstractGene — Type.

An abstraction layer to hide the implementation details of genes.

source

GeMM.BigGene — Type.

A variation of the standard Gene struct that allows for longer base sequence lengths through its use of BigInt.

source

GeMM.Chromosome — Type.

A struct that saves a collection of genes and a tag to denote whether they are maternal or not.

source

GeMM.Gene — Type.

The standard gene struct; including the base sequence (represented in binary and stored and as an integer), and an array of traits the gene codes for.

source

GeMM.Individual — Type.

One of the core structs of the model, representing an individual organism.

source

GeMM.Trait — Type.

A simple struct to store organism trait names and values.

source

initialisation.jl

submodels.jl

GeMM.disperse! — Function.

disperse!(w)

Dispersal of individuals within world (array of patches) w

source

GeMM.disturb! — Method.

disturb!(p,i)

species-independent mortality due to disturbance on patch p

source

GeMM.establish! — Function.

establish!(p, n)

establishment of individuals in patch p. Sets adaption parameter according to adaptation to number n niches of the surrounding environment.

source

GeMM.grow! — Method.

grow!(p)

Growth of individuals in patch p

source

GeMM.reproduce! — Method.

reproduce!(p, settings)

Reproduction of individuals in a patch p

source

GeMM.survive! — Method.

survive!(p)

density independent survival of individuals in patch p

source

scheduling.jl