Biological Processes

The files and functions listed below represent the different submodels of the model that implement its actual logic. They cover all the important physiological, life-history, ecological, and evolutionary processes simulated by GeMM.

genetics.jl

GeMM.createchrms — Method

createchrms(nchrms, genes)

Randomly distribute the passed genes into the given number of chromosomes. Returns an array of chromosome objects.

GeMM.creategenes — Method

creategenes(ngenes, traits, settings)

Randomly create a given number of gene objects, with their base sequence and associated traits. Returns the result as an array of AbstractGenes.

GeMM.createtraits — Method

createtraits(settings)

Create an array of trait objects generated from the default trait values (with a random offset).

GeMM.getmeantraitvalue — Method

getmeantraitvalue(traits, traitidx)

Take an array of traits and return the mean value of the indexed trait.

GeMM.getseq — Method

getseq(genome, traitidx)

Find and return the sequence of one gene that codes for the given trait traitidx.

GeMM.getseqsimilarity — Method

getseqsimilarity(seqone, seqtwo)

Compare two strings and return similarity.

GeMM.getstdtraitvalue — Method

getstdtraitvalue(traits, traitidx)

Take an array of traits and return the standard deviation of the indexed trait.

GeMM.gettraitdict — Method

gettraitdict(genes, traitnames)

Calculate the trait dict for an array of genes.

GeMM.gettraitdict — Method

gettraitdict(chromosomes, traitnames)

Convert a genome (an array of chromosomes) into a dict of traits and their values.

GeMM.gettraitdict — Method

gettraitdict(traits, traitnames)

Construct a trait dict from a list of Trait objects.

GeMM.iscompatible — Method

iscompatible(mate, individual, traitnames)

Check to see whether two individual organisms are reproductively compatible.

GeMM.meiosis — Method

meiosis(genome, maternal)

Carry out meiosis on a genome (marked as maternal or not). Returns a haploid gamete genome. (genome => array of chromosomes)

GeMM.num2seq — Method

num2seq(n)

Convert an integer into binary and then into a DNA base sequence string.

GeMM.seq2bignum — Method

seq2bignum(sequence)

Convert a DNA base sequence (a string) into binary and then into an BigInt (for larger genes). This saves memory.

GeMM.seq2num — Method

seq2num(sequence)

Convert a DNA base sequence (a string) into binary and then into an integer. This saves memory.

reproduction.jl

GeMM.createoffspring — Function

createoffspring(noffs, individual, partner, traitnames, dimorphism)

The main reproduction function. Take two organisms and create the given number of offspring individuals. Returns an array of individuals.

GeMM.findmate — Method

findmate(population, individual, traitnames)

Find a reproduction partner for the given individual in the given population.

GeMM.greproduce! — Method

greproduce!(patch, settings)

Reproduction of individuals in a patch with global mating.

GeMM.mutate! — Function

mutate!(traits, settings, locivar)

Loop over an array of traits, mutating each value in place along a normal distribution. locivar can be used to scale the variance of the normal distribution used to draw new trait values (together with settings[phylconstr]).

GeMM.mutate! — Method

mutate!(world, settings)

Mutate the world. (That sounds scary!)

GeMM.mutate! — Method

mutate!(individual, temp, settings)

Mutate an individual's genome (sequence and traits) in place.

GeMM.mutate! — Method

mutate!(patch, setting)

Mutate all seed individuals in a patch.

GeMM.reproduce! — Method

reproduce!(world, settings)

Carry out reproduction on all patches.

GeMM.reproduce! — Method

reproduce!(patch, settings)

Reproduction of individuals in a patch (default function).

survival.jl

GeMM.compete! — Function

compete!(world, static)

Carry out competition on all patches.

GeMM.compete! — Method

compete!(patch)

While there are too many organisms in a patch, pick two at random and kill the one that is less adapted to the local precipitation levels.

GeMM.grow! — Method

grow!(world, settings)

Carry out growth for all patches.

GeMM.grow! — Method

grow!(patch, growthrate, capgrowth)

Growth of individuals in the given patch. The actual growthrate is calculated with a metabolic formula, modified by the passed growthrate variable.

GeMM.survive! — Method

survive!(world, settings)

World-wide mortality. Sounds apocalyptic, but is just a fact of life.

GeMM.survive! — Method

survive!(patch, mortality)

Density independent survival of individuals in a patch. The actual mortality probability is calculated with a metabolic formula, modified by the passed mortality variable and an individual's temperature adaptation.

dispersal.jl

GeMM.checkborderconditions — Method

checkborderconditions!(w, x, y)

check if coordinates x and y lie within world w and correct if not, considering defined border conditions.

GeMM.checkviability! — Method

checkviability!(community, settings)

Check whether all individuals in the passed community conform to a basic set of constraints (i.e. all traits are present and certain properties are >= 0). Individuals that fail the test are removed from the community.

GeMM.checkviability! — Method

checkviability(world, settings)

Check the viability of all individuals.

GeMM.disperse! — Function

disperse!(world, static)

Dispersal of individuals within the world.

GeMM.establish! — Function

establish!(world, nniches, static)

Carry out establishment for each patch in the world.

GeMM.establish! — Function

establish!(patch, nniches)

Establishment of individuals in patch p: Sets the adaptation parameters (~fitness) according to an individual's adaptation to the niches of the surrounding environment.

A maximum of two niches (temperature and "precipitation") is currently supported.

GeMM.findisland — Method

findisland(w)

within world w, find out in which direction from the continent the island(s) lie(s).

GeMM.gausscurve — Function

gausscurve(b, c, x, a=1.0)

Calculate the value of the Gauss function ("bell curve") at point x; with a being the maximum height of the curve, b the position of the curve center and c the standard deviation ("width").

GeMM.traitsexist — Method

traitsexist(traits, settings)

Check a trait dict to make sure it contains the full set of traitnames required by the model (as defined in the settings).

GeMM.traitsexist — Method

traitsexist(individual, settings)

Make sure an individual organism has the full set of traits required by the model (as defined in the settings).