Internal functions

Internal functions

auxfuncts.jl

GeMM.checkborderconditions — Method.

checkborderconditions!(w, x, y)

check if coordinates x and y lie within world w and correct if not, considering defined border conditions.

source

GeMM.diversity — Method.

diversity(w)

Calculate (average) alpha, beta and gamma diversity of the world. Returns a tuple with the three values (a,b,g). cf. Veech et al. 2002

source

GeMM.findisland — Method.

findisland(w)

within world w, find out in which direction from the continent the island(s) lie(s).

source

GeMM.num2seq — Method.

num2seq(n)

Convert an integer into binary and then into a DNA base sequence string.

source

GeMM.seq2num — Method.

seq2num(sequence)

Convert a DNA base sequence (a string) into binary and then into an integer. This saves memory.

source