Internal functions

These are functions that the model uses in the background. They include a range of things: sub-submodels such findposspartner, utility functions like gausscurve, or performance hacks like seq2num.

auxfuncts.jl

GeMM.checkborderconditions — Method

checkborderconditions!(w, x, y)

check if coordinates x and y lie within world w and correct if not, considering defined border conditions.

GeMM.createchrms — Method

createchrms(nchrms, genes)

Randomly distribute the passed genes into the given number of chromosomes. Returns an array of chromosome objects.

GeMM.creategenes — Method

creategenes(ngenes, traits, settings)

Randomly create a given number of gene objects, with their base sequence and associated traits. Returns the result as an array of AbstractGenes.

GeMM.createind — Function

createind(settings, marked=false)

Create an individual organism with of a new species with a random genome.

GeMM.createoffspring — Method

createoffspring(noffs, individual, partner, traitnames)

The main reproduction function. Take two organisms and create the given number of offspring individuals. Returns an array of individuals.

GeMM.createtraits — Method

createtraits(settings)

Create an array of trait objects generated from the default trait values (with a random offset).

GeMM.diversity — Method

diversity(w)

Calculate (average) alpha, beta and gamma diversity of the world. Returns a tuple with the three values (a,b,g). cf. Veech et al. 2002

GeMM.findisland — Method

findisland(w)

within world w, find out in which direction from the continent the island(s) lie(s).

GeMM.findmate — Method

findmate(population, individual, traitnames)

Find a reproduction partner for the given individual in the given population.

GeMM.findmate — Method

findmate(patch, individual, traitnames)

Find a reproduction partner for the given individual in the given patch.

GeMM.freespace — Method

freespace(world)

Calculate the average amount of free space on each patch.

GeMM.gausscurve — Function

gausscurve(b, c, x, a=1.0)

Calculate the value of the Gauss function ("bell curve") at point x; with a being the maximum height of the curve, b the position of the curve center and c the standard deviation ("width").

GeMM.getmeantraitvalue — Method

getmeantraitvalue(traits, traitidx)

Take an array of traits and return the mean value of the indexed trait.

GeMM.getseq — Method

getseq(genome, traitidx)

Find and return the sequence of one gene that codes for the given trait traitidx.

GeMM.getseqsimilarity — Method

getseqsimilarity(seqone, seqtwo)

Compare two strings and return similarity.

GeMM.getstdtraitvalue — Method

getstdtraitvalue(traits, traitidx)

Take an array of traits and return the standard deviation of the indexed trait.

GeMM.gettraitdict — Method

gettraitdict(genes, traitnames)

Calculate the trait dict for an array of genes.

GeMM.gettraitdict — Method

gettraitdict(chromosomes, traitnames)

Convert a genome (an array of chromosomes) into a dict of traits and their values.

GeMM.gettraitdict — Method

gettraitdict(traits, traitnames)

Construct a trait dict from a list of Trait objects.

GeMM.identifyAdults! — Method

identifyAdults!(patch)

Build up the whoiswho index of individuals and species in a patch.

GeMM.iscompatible — Method

iscompatible(mate, individual, traitnames)

Check to see whether two individual organisms are reproductively compatible.

GeMM.markthem! — Method

markthem!(community)

Set each individual in the community (= array of individuals) as "marked".

GeMM.markthem! — Method

markthem!(world)

Set every individual in the world as "marked".

GeMM.markthem! — Method

markthem!(habitat)

Set each individual in the given patch as "marked".

GeMM.meiosis — Method

meiosis(genome, maternal)

Carry out meiosis on a genome (marked as maternal or not). Returns a haploid gamete genome. (genome => array of chromosomes)

GeMM.num2seq — Method

num2seq(n)

Convert an integer into binary and then into a DNA base sequence string.

GeMM.seq2bignum — Method

seq2bignum(sequence)

Convert a DNA base sequence (a string) into binary and then into an BigInt (for larger genes). This saves memory.

GeMM.seq2num — Method

seq2num(sequence)

Convert a DNA base sequence (a string) into binary and then into an integer. This saves memory.

GeMM.traitsexist — Method

traitsexist(traits, settings)

Check a trait dict to make sure it contains the full set of traitnames required by the model (as defined in the settings).

GeMM.traitsexist — Method

traitsexist(individual, settings)

Make sure an individual organism has the full set of traits required by the model (as defined in the settings).

GeMM.varyalleles! — Method

varyalleles!(genes, settings, locivar)

Mutate gene traits in the passed array of genes.

GeMM.varyalleles! — Method

varyalleles!(chromosomes, settings, locivar)

Mutate gene traits in the passed array of chromosomes.