Extensions

Originally, GeMM was designed to model island plant communities, but it has since been adapted to various other study questions. These files include functions that are not needed by the "core" model, but were added for these other studies.

habitatchange.jl

Investigate the effect of temporal habitat change on plant communities.

GeMM.changehabitat! — Method

changehabitat!(world)

Carry out 'global change' on all patches.

GeMM.changeprec! — Method

changeprec!(world, sdprec)

Change the precipitation of all patches according to a normal distribution.

GeMM.changetemp! — Method

changetemp!(world, sdtemp)

Change the temperature of all patches according to a normal distribution.

GeMM.markthem! — Method

markthem!(community)

Set each individual in the community (= array of individuals) as "marked".

GeMM.markthem! — Method

markthem!(world)

Set every individual in the world as "marked".

GeMM.markthem! — Method

markthem!(habitat)

Set each individual in the given patch as "marked".

invasion.jl

Carry out species invasions experiments in an island setting.

FIXME Documenter.jl doesn't recognise function definitions that are not on the top level - like the invade!() functions.

GeMM.disturb! — Method

disturb!(world)

Disturb all patches in the world.

GeMM.disturb! — Method

disturb!(patch, intensity)

Species-independent mortality due to disturbance on the given patch. The intensity gives the mortality percentage.

zosterops.jl

Adapt GeMM to investigate eco-evolutionary population dynamics of East-African Zosterops populations in the Taita Hills, Kenya.

FIXME Documenter.jl doesn't recognise function definitions that are not on the top level. This includes:

initzosteropsspecies()

getzosteropsspecies()

coordinate()

GeMM.findneighbours! — Method

findneighbours(world)

Construct a list of neighbours for each patch in the world, for faster lookup later on. (Must be called during initialisation.)

GeMM.zdisperse! — Method

zdisperse!(world)

Disperse all juvenile individuals within the world.

GeMM.zdisperse! — Method

zdisperse!(bird, patch, world)

Dispersal of a single bird. Birds look patches with a suitable habitat and a free territory or available mate. (Cf. Aben et al. 2016)

GeMM.zfindmate — Method

zfindmate(population, bird)

Find an available breeding partner in the given population, preferring conspecifics. Returns the partner individual, or nothing.

GeMM.zgenesis — Method

zgenesis(patch)

Create a new community of Zosterops breeding pairs (possibly of multiple species). Returns an array of individuals.

GeMM.ziscompatible — Method

ziscompatible(individual1, individual2)

Check to see whether two birds are reproductively compatible.

GeMM.zreproduce! — Method

zreproduce!(patch)

Reproduction of Zosterops breeding pairs in a patch.