Input, Output, and Settings

Input, Output, and Settings

These functions are responsible for reading in all model configurations (passed by config file or commandline), administrating them during a run, and printing any output.

constants.jl

defaults.jl

input.jl

GeMM.basicparser — Method.

basicparser(f)

Do elementary parsing on a config or map file.

Reads in the file, strips whole-line and inline comments and separates lines by whitespace. Returns a 2d array representing the tokens in each line.

source

output.jl

GeMM.setupdatadir — Method.

setupdatadir(dir)

Creates the output directory and copies relevant files into it. If the output directory already includes files, create a new directory by appending a counter.

source

GeMM.writedata — Method.

writedata(world, settings, timestep)

writes simulation output from world to separate table and fasta files. timestep and setting information is used for file name creation.

source

GeMM.dumpinds — Function.

dumpinds(world, io, sep)

Output all data of individuals in world as table to io. Columns are separated by sep.

source

GeMM.recordlineages — Method.

recordlineages(w)

Save the abundance of each lineage per patch

source

GeMM.recordstatistics — Method.

recordstatistics(w)

Write out world properties to the log file for later analysis.

source

GeMM.simlog — Function.

simlog(msg, settings, category)

Write a log message to STDOUT/STDERR and the specified logfile (if logging is turned on in the settings). Categories: d (debug), i (information, default), w (warn), e (error)

source