Main Model

These are the core functions of the model, implementing the high-level processes of the simulation and the object classes that are involved.

entities.jl

GeMM.AbstractGene — Type

An abstraction layer to hide the implementation details of genes.

GeMM.BigGene — Type

A variation of the standard Gene struct that allows for longer base sequence lengths through its use of BigInt.

GeMM.Chromosome — Type

A struct that saves a collection of genes and a tag to denote whether they are maternal or not.

GeMM.Gene — Type

The standard gene struct; including the base sequence (represented in binary and stored and as an integer), and an array of traits the gene codes for.

GeMM.Individual — Type

One of the core structs of the model, representing an individual organism.

GeMM.Patch — Type

One of the core structs of the model, representing a one-hectare patch of ground.

GeMM.Trait — Type

A simple struct to store organism trait names and values.

initialisation.jl

GeMM.createpop — Method

createpop(settings)

Create a new, random individual and replicates it a certain number of times (depending on metabolic variables) to create a new population of organisms. Returns an array of individuals.

FIXME Why does this not interface with auxfuncts/createind()?

GeMM.createworld — Method

createworld(maptable, settings)

Use a parsed map file (as returned by readmapfile) to create the world. Initialises each patch with its parameters and a new community, then returns an array of patches.

GeMM.genesis — Method

genesis(settings)

Create a new community, composed of random new species populations, for a patch. Returns an array of individuals.

GeMM.updateworld! — Method

updateworld!(world, maptable, cellsize)

Reinitialise the world from another parsed map file. Works analogously to createworld. Intended for use in scenarios where the model world changes during a run (e.g. through global warming or island ontogeny).

submodels.jl

GeMM.changehabitat! — Method

changehabitat!(world, settings)

Carry out 'global change' on all patches.

GeMM.changeprec! — Method

changeprec!(world, sdprec)

Change the precipitation of all patches according to a normal distribution.

GeMM.changetemp! — Method

changetemp!(world, sdtemp)

Change the temperature of all patches according to a normal distribution.

GeMM.checkviability! — Method

checkviability!(community, settings)

Check whether all individuals in the passed community conform to a basic set of constraints (i.e. all traits are present and certain properties are >= 0). Individuals that fail the test are removed from the community.

GeMM.checkviability! — Method

checkviability(world, settings)

Check the viability of all individuals.

GeMM.checkviability! — Method

checkviability!(patch, settings)

Check the viability of the individuals in this patch.

GeMM.compete! — Function

compete!(world, static)

Carry out competition on all patches.

GeMM.compete! — Method

compete!(patch)

While there are too many organisms in a patch, pick two at random and kill the one that is less adapted to the local precipitation levels.

GeMM.disperse! — Function

disperse!(world, static)

Dispersal of individuals within the world.

GeMM.disturb! — Method

disturb!(world, settings)

Disturb all patches in the world.

GeMM.disturb! — Method

disturb!(patch, intensity)

Species-independent mortality due to disturbance on the given patch. The intensity gives the mortality percentage.

GeMM.establish! — Function

establish!(patch, nniches)

Establishment of individuals in patch p: Sets the adaptation parameters (~fitness) according to an individual's adaptation to the niches of the surrounding environment.

A maximum of two niches is currently supported.

GeMM.establish! — Function

establish!(world, nniches, static)

Carry out establishment for each patch in the world.

GeMM.grow! — Method

grow!(world, settings)

Carry out growth for all patches.

GeMM.grow! — Method

grow!(patch, growthrate)

Growth of individuals in the given patch. The actual growthrate is calculated with a metabolic formula, modified by the passed growthrate variable.

GeMM.mutate! — Function

mutate!(traits, settings, locivar)

Loop over an array of traits, mutating each value in place along a normal distribution. locivar can be used to scale the variance of the normal distribution used to draw new trait values (together with settings[phylconstr]).

GeMM.mutate! — Method

mutate!(world, settings)

Mutate the world. (That sounds scary!)

GeMM.mutate! — Method

mutate!(individual, temp, settings)

Mutate an individual's genome (sequence and traits) in place.

GeMM.mutate! — Method

mutate!(patch, setting)

Mutate all seed individuals in a patch.

GeMM.reproduce! — Method

reproduce!(world, settings)

Carry out reproduction on all patches.

GeMM.reproduce! — Method

reproduce!(patch, settings)

Reproduction of individuals in a patch.

GeMM.survive! — Method

survive!(world, settings)

World-wide mortality. Sounds apocalyptic, but is just a fact of life.

GeMM.survive! — Method

survive!(patch, mortality)

Density independent survival of individuals in a patch. The actual mortality probability is calculated with a metabolic formula, modified by the passed mortality variable and an individual's temperature adaptation.

scheduling.jl

GeMM.simulate! — Function

simulate!(world, settings, timesteps)

This is the central function of the model with the main event loop. It defines the scheduling for all submodels and output functions.