Main Model

Main Model

These are the core functions of the model, implementing the high-level processes of the simulation and the object classes that are involved.

entities.jl

GeMM.Patch — Type.

One of the core structs of the model, representing a one-hectare patch of ground.

source

GeMM.AbstractGene — Type.

An abstraction layer to hide the implementation details of genes.

source

GeMM.BigGene — Type.

A variation of the standard Gene struct that allows for longer base sequence lengths through its use of BigInt.

source

GeMM.Chromosome — Type.

A struct that saves a collection of genes and a tag to denote whether they are maternal or not.

source

GeMM.Gene — Type.

The standard gene struct; including the base sequence (represented in binary and stored and as an integer), and an array of traits the gene codes for.

source

GeMM.Individual — Type.

One of the core structs of the model, representing an individual organism.

source

GeMM.Trait — Type.

A simple struct to store organism trait names and values.

source

initialisation.jl

GeMM.createworld — Method.

createworld(maptable, settings)

Use a parsed map file (as returned by readmapfile) to create the world. Initialises each patch with its parameters and a new community, then returns an array of patches.

source

GeMM.updateworld! — Method.

updateworld!(world, maptable, cellsize)

Reinitialise the world from another parsed map file. Works analogously to createworld. Intended for use in scenarios where the model world changes during a run (e.g. through global warming or island growth).

source

GeMM.createpop — Method.

createpop(settings)

Create a new, random individual and replicates it a certain number of times (depending on metabolic variables) to create a new population of organisms. Returns an array of individuals.

XXX Why does this not interface with auxfuncts/createind()?

source

GeMM.genesis — Method.

genesis(settings)

Create a new community, composed of random new species populations, for a patch. Returns an array of individuals.

source

submodels.jl

GeMM.changehabitat! — Method.

changehabitat!(world, settings)

Carry out 'global change' on all patches.

source

GeMM.changeprec! — Method.

changeprec!(world, sdprec)

Change the precipitation of all patches according to a normal distribution.

source

GeMM.changetemp! — Method.

changetemp!(world, sdtemp)

Change the temperature of all patches according to a normal distribution.

source

GeMM.checkviability! — Method.

checkviability!(community, settings)

Check whether all individuals in the passed community conform to a basic set of constraints (i.e. all traits are present and certain properties are >= 0). Individuals that fail the test are removed from the community.

source

GeMM.checkviability! — Method.

checkviability(world, settings)

Check the viability of all individuals.

source

GeMM.checkviability! — Method.

checkviability!(patch, settings)

Check the viability of the individuals in this patch.

source

GeMM.compete! — Function.

compete!(world, static)

Carry out competition on all patches.

source

GeMM.compete! — Method.

compete!(patch)

While there are too many organisms in a patch, pick two at random and kill the one that is less adapted to the local precipitation levels.

source

GeMM.disperse! — Function.

disperse!(world, static)

Dispersal of individuals within the world.

source

GeMM.disturb! — Method.

disturb!(world, settings)

Disturb all patches in the world.

source

GeMM.disturb! — Method.

disturb!(patch, intensity)

Species-independent mortality due to disturbance on the given patch. The intensity gives the mortality percentage.

source

GeMM.establish! — Function.

establish!(world, nniches, static)

Carry out establishment for each patch in the world.

source

GeMM.establish! — Function.

establish!(patch, nniches)

Establishment of individuals in patch p: Sets the adaption parameters (~fitness) according to an individual's adaptation to the niches of the surrounding environment.

A maximum of two niches is currently supported.

source

GeMM.grow! — Method.

grow!(world, settings)

Carry out growth for all patches.

source

GeMM.grow! — Method.

grow!(patch, growthrate)

Growth of individuals in the given patch. The actual growthrate is calculated with a metabolic formula, modified by the passed growthrate variable.

source

GeMM.mutate! — Function.

mutate!(traits, settings, locivar)

Loop over an array of traits, mutating each value in place along a normal distribution.

XXX what does locivar do?

source

GeMM.mutate! — Method.

mutate!(world, settings)

Mutate the world. (That sounds scary!)

source

GeMM.mutate! — Method.

mutate!(individual, temp, settings)

Mutate an individual's genome (sequence and traits) in place.

source

GeMM.mutate! — Method.

mutate!(patch, setting)

Mutate all seedling individual's in a patch.

source

GeMM.reproduce! — Method.

reproduce!(world, settings)

Carry out reproduction on all patches.

source

GeMM.reproduce! — Method.

reproduce!(patch, settings)

Reproduction of individuals in a patch.

source

GeMM.survive! — Method.

survive!(world, settings)

World-wide mortality. Sounds apocalyptic, but is just a fact of life.

source

GeMM.survive! — Method.

survive!(patch, mortality)

Density independent survival of individuals in a patch. The actual mortality probability is calculated with a metabolic formula, modified by the passed mortality variable and an individual's temperature adaptation.

source

scheduling.jl

GeMM.simulate! — Function.

simulate!(world, settings, timesteps)

This is the central function of the model with the main event loop. It defines the scheduling for all submodels and output functions.

source