Main Model

Main Model

These are the core functions of the model, implementing the high-level processes of the simulation and the object classes that are involved.

entities.jl

initialisation.jl

submodels.jl

GeMM.disperse! — Function.

disperse!(w)

Dispersal of individuals within world (array of patches) w

source

GeMM.disturb! — Method.

disturb!(p,i)

species-independent mortality due to disturbance on patch p

source

GeMM.establish! — Function.

establish!(p, n)

establishment of individuals in patch p. Sets adaption parameter according to adaptation to number n niches of the surrounding environment.

source

GeMM.grow! — Method.

grow!(p)

Growth of individuals in patch p

source

GeMM.reproduce! — Method.

reproduce!(p, settings)

Reproduction of individuals in a patch p

source

GeMM.survive! — Method.

survive!(p)

density independent survival of individuals in patch p

source

scheduling.jl