Biological Processes

The files and functions listed below represent the different submodels of the model that implement its actual logic. They cover all the important physiological, life-history, ecological, and evolutionary processes simulated by GeMM.

genetics.jl

GeMM.createchrms — Method

createchrms(nchrms, genes)

Randomly distribute the passed genes into the given number of haploid chromosomes. Returns a diploid genome (array of chromosome objects).

GeMM.createcompatgene — Method

createcompatgene(bases, compatidx)

Create the gene that is used to determine reproductive compatibility.

GeMM.creategenes — Method

creategenes(ngenes, traits)

Randomly create a given number of gene objects, with their base sequence and associated traits. Returns the result as an array of AbstractGenes.

GeMM.createtraits — Method

createtraits()

Create an array of trait objects generated from the default trait values (with a random offset).

GeMM.getseqs — Method

getseqs(genome, traitname)

Find and return the sequences of genes that code for the given trait.

GeMM.getseqsimilarity — Method

getseqsimilarity(seqone, seqtwo)

Compare two strings and return similarity.

GeMM.gettraitdict — Method

gettraitdict(chromosomes, traitnames)

Convert a genome (an array of chromosomes) into a dict of traits and their values.

GeMM.gettraitdictfast — Method

gettraitdictfast(chromosomes, traitnames)

Convert a genome (an array of chromosomes) into a dict of traits and their values. This is an optimised version that can be run if degpleiotropy is 0 and linkage is "none".

GeMM.iscompatible — Method

iscompatible(mate, individual)

Check to see whether two individual organisms are reproductively compatible by comparing gene sequences (which sequence depends on setting("speciation")).

GeMM.meiosis — Method

meiosis(genome, maternal)

Carry out meiosis on a genome (marked as maternal or not). Returns a haploid gamete genome. (genome => array of chromosomes)

GeMM.mutate! — Function

mutate!(traits, locivar)

Loop over an array of traits, mutating each value in place along a normal distribution. locivar can be used to scale the variance of the normal distribution used to draw new trait values (together with setting(phylconstr]).

GeMM.mutate! — Method

mutate!(world)

Mutate the world. (That sounds scary!)

GeMM.mutate! — Method

mutate!(individual, temp)

Mutate an individual's genome (sequence and traits) in place.

GeMM.mutate! — Method

mutate!(patch)

Mutate all seed individuals in a patch.

GeMM.num2seq — Method

num2seq(n)

Convert an integer into binary and then into a DNA base sequence string.

GeMM.seq2bignum — Method

seq2bignum(sequence)

Convert a DNA base sequence (a string) into binary and then into a BigInt (for larger genes). This saves memory.

GeMM.varyalleles! — Method

varyalleles!(genes, locivar)

Mutate gene traits in the passed array of genes.

GeMM.varyalleles! — Method

varyalleles!(chromosomes, locivar)

Mutate gene traits in the passed array of chromosomes.

reproduction.jl

GeMM.createoffspring — Function

createoffspring(noffs, individual, partner, dimorphism)

The main reproduction function. Take two organisms and create the given number of offspring individuals. Returns an array of individuals.

GeMM.findmate — Method

findmate(population, individual)

Find a reproduction partner for the given individual in the given population.

GeMM.greproduce! — Method

greproduce!(patch)

Reproduction of individuals in a patch with global mating.

GeMM.reproduce! — Method

reproduce!(world)

Carry out reproduction on all patches.

GeMM.reproduce! — Method

reproduce!(patch)

Reproduction of individuals in a patch (default function).

survival.jl

GeMM.compete! — Method

compete!(world)

Carry out competition on all patches.

GeMM.compete! — Method

compete!(patch)

While there are too many organisms in a patch, pick two at random and kill the one that is less adapted to the local precipitation levels.

GeMM.grow! — Method

grow!(world)

Carry out growth for all patches.

GeMM.grow! — Method

grow!(patch, growthrate, capgrowth)

Growth of individuals in the given patch. The actual growthrate is calculated with a metabolic formula, modified by the passed growthrate variable.

GeMM.survive! — Method

survive!(world)

World-wide mortality. Sounds apocalyptic, but is just a fact of life.

GeMM.survive! — Method

survive!(patch)

Density independent survival of individuals in a patch. The actual mortality probability is calculated with a metabolic formula, modified by the passed mortality variable and an individual's temperature adaptation.

dispersal.jl

GeMM.checkborderconditions — Method

checkborderconditions!(w, x, y)

Check if the coordinates x and y lie within world w and correct if not, considering the user-defined border conditions.

GeMM.checkviability! — Method

checkviability(world)

Check the viability of all individuals. (Note: this does not have to be called every update, because viability checking is already integrated into establish!(). However, it may be useful to call once after initialisation.)

GeMM.disperse! — Method

disperse!(world, static)

Dispersal of individuals within the world.

GeMM.establish! — Method

establish!(world)

Carry out establishment for each patch in the world.

GeMM.establish! — Method

establish!(patch)

Establishment of individuals in patch p: Sets the adaptation parameters (~fitness) according to an individual's adaptation to the niches of the surrounding environment. Also tests whether the individual is viable.

GeMM.findisland — Method

findisland(w)

Within world w, find out in which direction from the continent the island lies.

GeMM.gausscurve — Function

gausscurve(b, c, x, a=1.0)

Calculate the value of the Gauss function ("bell curve") at point x; with a being the maximum height of the curve, b the position of the curve center and c the standard deviation ("width").

GeMM.isviable — Method

isviable(individual)

Check whether this individual conforms to a basic set of constraints (i.e. all traits are present and certain properties are >= 0).

GeMM.traitsexist — Method

traitsexist(traits)

Check a trait dict to make sure it contains the full set of traitnames required by the model (as defined in the settings).